[week4-task1] util을 lib 디렉토리로 옮기기
