--- src/utils.js
@@ -1,3 +1,0 @@
-export function get(key) {
-  return (obj) => obj[key];
-}
