--- src/utils.test.js
@@ -1,13 +1,0 @@
-import { get, equal } from './utils';
-
-test('get', () => {
-  const state = {
-    name: '홍길동',
-  };
-
-  const f = get('name');
-  const g = get('age');
-
-  expect(f(state)).toBe('홍길동');
-  expect(g(state)).toBeUndefined();
-});
